Move shed inside commands.
